opkg: introduce a --force-checksum cmdline flag to be ale to ignore mismatching md5sums

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@40766 3c298f89-4303-0410-b956-a3cf2f4a3e73

The patch was imported from OpenWRT git server
(git://git.openwrt.org/openwrt.git) as of commit id
37a0e124d1c685f0503b7b96351e29548fbe71cc.

Parts have been re-written to match opkg version 0.3.x:
 libopkg/opkg_install.c is no longer patched; error messages for md5sum
 mismatch no longer exist in this version of opkg. Patching libopkg/pkg.c
 instead, for ignoring checksum mismatch (new code) patching information
 for src/opkg-cl.c (version 0.1.8) used now to patch src/opkg.c (the new
 program source)
 man/opkg.1.in - add description for the "--force-checksum" command line switch.
 man/opkg.5.in - add description for the "force_checksum" config option.

+++ libopkg/pkg.c
@@ -1444,8 +1444,17 @@ int pkg_verify(pkg_t * pkg)
 
  fail:
     free(local_sig_filename);
-    opkg_msg(NOTICE, "Removing corrupt package file %s.\n",
+    if (!opkg_config->force_checksum)
+    {
+	opkg_msg(NOTICE, "Removing corrupt package file %s.\n",
              pkg->local_filename);
-    unlink(pkg->local_filename);
-    return err;
+	unlink(pkg->local_filename);
+	return err;
+    }
+    else
+    {
+	opkg_msg(NOTICE, "Ignored %s checksum mismatch.\n",
+             pkg->local_filename);
+        return 0;
+    }
 }
